Ich habe das Projekt erweitert (#1)
* Add files via upload

This is just for testing, ignore

* Add files via upload

those are correct

* Added a class to make it easier to extend the states.
Every Class contains the path to the base of the bmp files. Bmp files have to be 240x240 and will be sent to the RaspberryPi

* Added a class to make it easier to extend the states.
Every Class contains the path to the base of the bmp files. Bmp files have to be 240x240 and will be sent to the RaspberryPi

* Delete testBmp directory

* Kotlin Code is now documented

* Kotlin Code is now documented

* C-Code added and README.md updated

---------

Co-authored-by: Valentin Werner <vwerner1@stud.hs-offenburg.de>